deletion only for uploader

+++ src/Film.svelte
@@ -2,13 +2,8 @@
 	import { deleteDoc,doc,increment,updateDoc } from "firebase/firestore";
 	import { getDownloadURL,ref } from "firebase/storage";
 	import { db,storage } from './firebase.js';
-	export let title;
-	export let directors;
-	export let cast;
-	export let synopsis;
-	export let release;
+	export let title, directors, cast, synopsis, release, id, currentUser, uploadId, likes, dislikes;
 	export let close = false;
-	export let id;
 
 	async function updateLikes(id, like){
   if(like > 0){
@@ -19,8 +14,11 @@
 }
 
 	async function deleteFilm(id){
-		await deleteDoc(doc(db, "films", id))
-		console.log(id)
+		let sure = confirm("Are you sure?");
+		if(sure){
+			await deleteDoc(doc(db, "films", id))
+		}
+		console.log(sure)
 	}
 
 	$: getDownloadURL(ref(storage, `images/${id}.jpeg`)).then((url) =>{
@@ -29,20 +27,32 @@
 	
 </script>
 <div id='film'>
-	<p>Id: {id}</p>
+{#if currentUser === uploadId}
+<div class="row">
+<div class="col-8"></div>
+<button class="btn btn-danger col-4 material-icons" id='deleteButton' on:click={()=>deleteFilm(id)}>delete_forever</button>
+</div>
+{/if}
 <img id="image" alt="Movie">
 <p>Title: {title}</p>
 <p>Release date: {release.toDate().getDate()}.{release.toDate().getMonth()}.{release.toDate().getFullYear()}</p>
 <p>Directors: {directors}</p>
 <p>Cast: {cast}</p>
-<button id='deleteButton' on:click={()=>deleteFilm(id)}>Delete!</button>
-<!-- <button on:click|once={()=>updateLikes(id,1)}>Like</button>
-<button on:click|once={()=>updateLikes(id,-1)}>Dislike</button> -->
+<div class="row">
+<button class="col-4 btn btn-primary material-icons" on:click|once={()=>updateLikes(id,1)}>thumb_up</button>
+<div class="col-2">Likes: {likes}</div>
+<div class="col-2">Dislikes: {dislikes}</div>
+<button class="col-4 btn btn-primary material-icons" on:click|once={()=>updateLikes(id,-1)}>thumb_down_alt</button>
+</div>
 <p>Synopsis: {synopsis}</p>
 </div>
 
 <style>
 	#image{
 		width: 100%;
+	}
+
+	#deleteButton{
+		right: 10;
 	}
 </style>
